+++ src/main/java/com/alife/tolsim/animals/ShadowIndividual.java
@@ -45,7 +45,7 @@ public class ShadowIndividual extends IndividualWithProperties {
      * @param date     in-simulation time
      * @param parent   parent id
      */
-    public ShadowIndividual(int glID, int ancestor, int date, int parent) {
+    public ShadowIndividual(long glID, long ancestor, int date, long parent) {
         super(glID, ancestor, date, parent);
     }
 
@@ -56,7 +56,7 @@ public class ShadowIndividual extends IndividualWithProperties {
      * @param in   individual to be cloned
      * @param glID id of this one
      */
-    public ShadowIndividual(ShadowIndividual in, int glID, int date, double cst_mut_factor, int cst_speed_max,
+    public ShadowIndividual(ShadowIndividual in, long glID, int date, double cst_mut_factor, int cst_speed_max,
                             double cst_light_birth_dst, double birth_dst, int cst_grid_max, int cst_energy_max
     ) {
         super(in, glID, date, cst_mut_factor, cst_speed_max, cst_light_birth_dst, birth_dst, cst_grid_max, cst_energy_max);
